matching-engine/requirements.txt
kafka-python
redis
requests
flask
flask_socketio
eventlet
flask-cors